Extracted the shared URL force-load helpers for input serializers
ref https://linear.app/ghost/issue/HKG-1822

forceUrlRelationsWhenLazy was duplicated byte-for-byte in the posts and
pages input serializers, and its "will this response carry a url" condition
was the hand-written inverse of the output serializer's guard — two spots
that must never drift, because a drift silently under-fetches and
reintroduces the thin-resource errors. The predicate now lives in
endpoints/utils as willSerializeUrl, used by both sides, and the force-load
moved next to its sibling forceUrlColumnsWhenLazy, parameterised by router
type.
