+++ ghost/core/core/server/api/endpoints/utils/index.js
@@ -49,5 +49,21 @@ module.exports = {
      */
     isPreview: (frame) => {
         return frame.isPreview === true;
+    },
+
+    /**
+     * @description Whether the response will carry a serialized `url` for the
+     * requested resources. True unless a `?fields` narrowing excludes it.
+     *
+     * Shared by the output serializer (which skips URL generation when false)
+     * and the input serializers (which force-load the columns/relations the
+     * URL service needs when true) — both sides must always agree, or the
+     * input under-fetches and the lazy URL service rejects the resource as
+     * thin.
+     * @param {import('@tryghost/api-framework').Frame} frame
+     * @returns {boolean}
+     */
+    willSerializeUrl: (frame) => {
+        return !Array.isArray(frame.options.columns) || frame.options.columns.includes('url');
     }
 };
