+++ ghost/core/core/server/api/endpoints/utils/serializers/input/pages.js
@@ -1,5 +1,4 @@
 const _ = require('lodash');
-const urlService = require('../../../../../services/url');
 const debug = require('@tryghost/debug')('api:endpoints:utils:serializers:input:pages');
 const {ValidationError} = require('@tryghost/errors');
 const tpl = require('@tryghost/tpl');
@@ -50,32 +49,6 @@ function selectAllAllowedColumns(frame) {
     }
 }
 
-function forceUrlRelationsWhenLazy(frame) {
-    // `url` is computed for every serialized page unless `?fields` narrows it
-    // away, so the relations the lazy URL service reads (e.g. tags for a
-    // tag-filtered collection) must be loaded whenever no columns were
-    // requested (plain browse/read) — not only for the `?fields=url` case.
-    const computesUrl = !Array.isArray(frame.options.columns) || frame.options.columns.includes('url');
-    if (!computesUrl) {
-        return;
-    }
-    const relations = urlService.facade.getRequiredRelations();
-    if (relations.length) {
-        const requested = frame.options.withRelated || [];
-        const forced = relations.filter(relation => !requested.includes(relation));
-        if (forced.length) {
-            // Record what was forced (not what the caller asked for) so the
-            // output mapper can strip it from the response after the URL is
-            // built — the response shape must match the query.
-            frame.forcedUrlRelations = forced;
-            frame.options.withRelated = _.union(requested, forced);
-        }
-    }
-    if (Array.isArray(frame.options.columns)) {
-        url.forceUrlColumnsWhenLazy(frame, 'pages');
-    }
-}
-
 function defaultRelations(frame) {
     // Defaults for the admin API. Applied before the URL force-load so a
     // forced relation can never preempt the full admin default list.
@@ -83,7 +56,7 @@ function defaultRelations(frame) {
         frame.options.withRelated = ['tags', 'authors', 'authors.roles', 'tiers', 'count.signups', 'count.paid_conversions'];
     }
 
-    forceUrlRelationsWhenLazy(frame);
+    url.forceUrlRelationsWhenLazy(frame, 'pages');
 }
 
 function setDefaultOrder(frame) {
@@ -153,7 +126,7 @@ module.exports = {
 
             setDefaultOrder(frame);
             forceVisibilityColumn(frame);
-            forceUrlRelationsWhenLazy(frame);
+            url.forceUrlRelationsWhenLazy(frame, 'pages');
         }
 
         if (!localUtils.isContentAPI(frame)) {
@@ -172,7 +145,7 @@ module.exports = {
             removeSourceFormats(frame);
             setDefaultOrder(frame);
             forceVisibilityColumn(frame);
-            forceUrlRelationsWhenLazy(frame);
+            url.forceUrlRelationsWhenLazy(frame, 'pages');
         }
 
         if (!localUtils.isContentAPI(frame)) {
